not merge this is a proposal
I think the idea of keeping everything in a view is very good because
it is more user friendly, however it is difficult to do tests,
I think it would be better to design, follow the CRUD convention.

I mention it to continue doing the tests

what do you think?