Improve Promote students feature
Fix handling promoting students for different year as session. Previously if students of a section are promoted then they could not be identified from the already existing students of that class section. Now promoted or new students will be labeled so that formar students of the class section can be promoted to next class.